consolidate style and some functions
