--- packages/fast-tooling-react/src/navigation/navigation.utilities.style.ts
@@ -1,69 +1,0 @@
-import { CSSRules } from "@microsoft/fast-jss-manager";
-import {
-    Direction,
-    ellipsis,
-    localizeSpacing,
-    toPx,
-} from "@microsoft/fast-jss-utilities";
-
-export interface BoxShadowConfig {
-    offsetX: number;
-    offsetY: number;
-    blurRadius: number;
-    spreadRadius: number;
-    color: string;
-    inset?: boolean;
-}
-
-export function localizePadding(
-    top: number,
-    right: number,
-    bottom: number,
-    left: number
-): CSSRules<{}> {
-    return {
-        padding: localizeSpacing(Direction.ltr)(
-            `${toPx(top)} ${toPx(right)} ${toPx(bottom)} ${toPx(left)}`
-        ),
-    };
-}
-
-export function applyTriggerStyle(color: string): CSSRules<{}> {
-    return {
-        lineHeight: "16px",
-        fontSize: "12px",
-        ...localizePadding(3, 5, 2, 16),
-        border: "none",
-        outline: "none",
-        boxSizing: "border-box",
-        color,
-        ...ellipsis(),
-    };
-}
-
-export function boxShadow(config: BoxShadowConfig): CSSRules<{}> {
-    return {
-        boxShadow: `${config.inset ? "inset " : ""}
-        ${toPx(config.offsetX)} ${toPx(config.offsetY)} ${toPx(config.blurRadius)} ${toPx(
-            config.spreadRadius
-        )} ${config.color}`,
-    };
-}
-
-/**
- * Mimics a border but with boxShadow (strong in the sense of no blur)
- */
-export function insetStrongBoxShadow(color: string): CSSRules<{}> {
-    const shadow: BoxShadowConfig = {
-        offsetX: 0,
-        offsetY: 0,
-        blurRadius: 0,
-        spreadRadius: 1,
-        color,
-        inset: true,
-    };
-
-    return {
-        ...boxShadow(shadow),
-    };
-}
