+++ packages/fast-tooling-react/src/style/index.ts
@@ -1,1 +1,2 @@
 export * from "./constants";
+export * from "./utilities";
