+++ packages/fast-tooling-react/src/style/utilities.ts
@@ -29,6 +29,19 @@ export function localizePadding(
     };
 }
 
+export function applyTriggerStyle(color: string): CSSRules<{}> {
+    return {
+        lineHeight: "16px",
+        fontSize: "12px",
+        ...localizePadding(3, 5, 2, 16),
+        border: "none",
+        outline: "none",
+        boxSizing: "border-box",
+        color,
+        ...ellipsis(),
+    };
+}
+
 /**
  * Used for styles radio buttons (vertical and horizontal alignment)
  */
